requirements.txt
langchain
langchain-google-genai
python-dotenv
google-auth-oauthlib
google-api-python-client
PyMuPDF